# swagger.yml
openapi: 3.0.0
info:
  title: Pediforte Student Management API
  description: API for managing student registrations, rules, and admin operations
  version: "1.0.0"
servers:
  - url: http://localhost:5000/api
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    StudentInformation:
      type: object
      properties:
        id:
          type: integer
        surname:
          type: string
        given_name:
          type: string
        other_names:
          type: string
        home_address:
          type: string
        phone_number:
          type: string
        email_address:
          type: string
        dob:
          type: string
          nullable: true
          example: ""
        gender:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        terms_agreed:
          type: boolean
        terms_agreed_at:
          type: string
          format: date-time
        course_info:
          $ref: '#/components/schemas/CourseInformation'
        payment_info:
          $ref: '#/components/schemas/PaymentInformation'
    CourseInformation:
      type: object
      properties:
        id:
          type: integer
        preferred_course:
          type: string
        objectives:
          type: array
          items:
            type: string
        prior_computer_knowledge:
          type: string
        seek_employment_opportunities:
          type: boolean
        hear_about_pediforte:
          type: string
        registration_date:
          type: string
          nullable: true
          example: ""
        resumption_date:
          type: string
          nullable: true
          example: ""
        created_at:
          type: string
          format: date-time
    PaymentInformation:
      type: object
      properties:
        id:
          type: integer
        course_price:
          type: number
        amount_paid:
          type: number
        balance:
          type: number
        payment_method:
          type: string
        receipt_no:
          type: string
        payment_status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    StudentAgreement:
      type: object
      properties:
        id:
          type: integer
        student_id:
          type: integer
        rules_id:
          type: integer
        rules_version:
          type: string
        agreed_at:
          type: string
          format: date-time
        ip_address:
          type: string
        user_agent:
          type: string
paths:
  /student-information:
    post:
      tags:
        - Student Management
      summary: Create a new student record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - surname
                - given_name
                - email_address
              properties:
                surname:
                  type: string
                  example: "string"
                given_name:
                  type: string
                  example: "string"
                other_names:
                  type: string
                  example: "string"
                home_address:
                  type: string
                  example: "string"
                phone_number:
                  type: string
                  example: "string"
                email_address:
                  type: string
                  example: "string"
                dob:
                  type: string
                  nullable: true
                  example: ""
                gender:
                  type: string
                  example: "string"
      responses:
        '201':
          description: Student created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentInformation'
        '400':
          description: Missing required fields or invalid data
  /course-information/{student_id}:
    post:
      tags:
        - Course Management
      summary: Create course information for a student
      parameters:
        - in: path
          name: student_id
          required: true
          schema:
            type: integer
          description: ID of the student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - preferred_course
                - objectives
              properties:
                preferred_course:
                  type: string
                  enum:
                    - Fullstack Development
                    - Frontend Development
                    - Cybersecurity
                    - Data Science
                    - Mobile App Development
                    - UI/UX Design
                  example: "string"
                objectives:
                  type: array
                  items:
                    type: string
                    enum:
                      - Expand My Knowledge
                      - Get a Job
                      - Competitive Edge
                      - Looking for Opportunities
                  example: ["string"]
                prior_computer_knowledge:
                  type: string
                  example: "string"
                seek_employment_opportunities:
                  type: boolean
                hear_about_pediforte:
                  type: string
                  example: "string"
                registration_date:
                  type: string
                  nullable: true
                  example: ""
                resumption_date:
                  type: string
                  nullable: true
                  example: ""
      responses:
        '201':
          description: Course information created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseInformation'
        '400':
          description: Invalid course, objectives, or data
        '404':
          description: Student not found
  /payment-information/{student_id}:
    post:
      tags:
        - Payment Management
      summary: Create payment information for a student
      parameters:
        - in: path
          name: student_id
          required: true
          schema:
            type: integer
          description: ID of the student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - course_price
                - payment_method
              properties:
                course_price:
                  type: number
                amount_paid:
                  type: number
                payment_method:
                  type: string
                  enum:
                    - cash
                    - bank_transfer
                receipt_no:
                  type: string
                  example: "string"
      responses:
        '201':
          description: Payment information created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentInformation'
        '400':
          description: Invalid payment method or data
        '404':
          description: Student not found
...
  /student-agreement/{student_id}:
    post:
      tags:
        - Student Rules
      summary: Record student agreement to rules
      parameters:
        - in: path
          name: student_id
          required: true
          schema:
            type: integer
          description: ID of the student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - agreed
              properties:
                agreed:
                  type: boolean
      responses:
        '201':
          description: Agreement recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentAgreement'
        '400':
          description: Agreement required or no active rules
        '404':
          description: Student not found
  /admin/students:
    get:
      tags:
        - Admin
      summary: List all students with optional filters
      security:
        - basicAuth: []
      parameters:
        - in: query
          name: gender
          schema:
            type: string
          description: Filter by gender
        - in: query
          name: preferred_course
          schema:
            type: string
          description: Filter by preferred course
        - in: query
          name: objectives
          schema:
            type: string
          description: Filter by objectives (comma-separated)
        - in: query
          name: payment_status
          schema:
            type: string
          description: Filter by payment status
        - in: query
          name: terms_agreed
          schema:
            type: boolean
          description: Filter by terms agreed status
      responses:
        '200':
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentInformation'
        '401':
          description: Unauthorized
  /admin/students/{student_id}:
    get:
      tags:
        - Admin
      summary: Get a specific student
      security:
        - basicAuth: []
      parameters:
        - in: path
          name: student_id
          required: true
          schema:
            type: integer
          description: ID of the student
      responses:
        '200':
          description: Student details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentInformation'
        '401':
          description: Unauthorized
        '404':
          description: Student not found
    put:
      tags:
        - Admin
      summary: Update a student’s information
      security:
        - basicAuth: []
      parameters:
        - in: path
          name: student_id
          required: true
          schema:
            type: integer
          description: ID of the student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                surname:
                  type: string
                  example: "string"
                given_name:
                  type: string
                  example: "string"
                other_names:
                  type: string
                  example: "string"
                home_address:
                  type: string
                  example: "string"
                phone_number:
                  type: string
                  example: "string"
                email_address:
                  type: string
                  example: "string"
                dob:
                  type: string
                  nullable: true
                  example: ""
                gender:
                  type: string
                  example: "string"
                preferred_course:
                  type: string
                  enum:
                    - Fullstack Development
                    - Frontend Development
                    - Cybersecurity
                    - Data Science
                    - Mobile App Development
                    - UI/UX Design
                  example: "string"
                objectives:
                  type: array
                  items:
                    type: string
                    enum:
                      - Expand My Knowledge
                      - Get a Job
                      - Competitive Edge
                      - Looking for Opportunities
                  example: ["string"]
                prior_computer_knowledge:
                  type: string
                  example: "string"
                seek_employment_opportunities:
                  type: boolean
                hear_about_pediforte:
                  type: string
                  example: "string"
                registration_date:
                  type: string
                  nullable: true
                  example: ""
                resumption_date:
                  type: string
                  nullable: true
                  example: ""
                course_price:
                  type: number
                amount_paid:
                  type: number
                payment_method:
                  type: string
                  enum:
                    - cash
                    - bank_transfer
                receipt_no:
                  type: string
                  example: "string"
      responses:
        '200':
          description: Student updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentInformation'
        '400':
          description: Invalid data
        '401':
          description: Unauthorized
        '404':
          description: Student not found
    delete:
      tags:
        - Admin
      summary: Delete a student
      security:
        - basicAuth: []
      parameters:
        - in: path
          name: student_id
          required: true
          schema:
            type: integer
          description: ID of the student
      responses:
        '200':
          description: Student deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Student not found
  /admin/students/report/csv:
    get:
      tags:
        - Admin
      summary: Generate CSV report of students
      security:
        - basicAuth: []
      parameters:
        - in: query
          name: gender
          schema:
            type: string
          description: Filter by gender
        - in: query
          name: preferred_course
          schema:
            type: string
          description: Filter by preferred course
        - in: query
          name: objectives
          schema:
            type: string
          description: Filter by objectives (comma-separated)
        - in: query
          name: payment_status
          schema:
            type: string
          description: Filter by payment status
        - in: query
          name: terms_agreed
          schema:
            type: boolean
          description: Filter by terms agreed status
      responses:
        '200':
          description: CSV report generated
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorized
...